maybe this worked this time?

+++ Assets/Scripts/EnemyController.cs
@@ -7,8 +7,7 @@ public class EnemyController : MonoBehaviour
     [SerializeField] private float fireRate;
     [SerializeField] private float bulletSpeedModifier;
     [SerializeField] private int startingHealth;
-    
-    private int health;
+    private float health;
     private PathController[] pathControllers;
     private Transform bulletSpawnPoint;
     private int currentMovementIdx = 0;
@@ -51,7 +50,7 @@ public class EnemyController : MonoBehaviour
         simpleBullet.SetSize(0.3f);
 
     }
-    public void Damage(int damage)
+    public void Damage(float damage)
     {
         health -= damage;
         if (health <= 0)
